Moved ObjectCache into an internal module.
This type has always been internal-only, but moving it to a module named `Internal` will make this clearer.

PiperOrigin-RevId: 606649281
